Move PublicationContext and Journals to package contexttypes
