+++ src/main/java/no/unit/nva/model/contexttypes/PublicationContext.java
@@ -1,7 +1,8 @@
-package no.unit.nva.model;
+package no.unit.nva.model.contexttypes;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
+import no.unit.nva.model.Level;
 
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type")
 @JsonSubTypes({@JsonSubTypes.Type(name = "Journal", value = Journal.class)})
